+++ stores/mysql/src/test/kotlin/com/gatehill/corebot/store/mysql/MysqlDataStoreImplSpec.kt
@@ -5,6 +5,7 @@ import com.gatehill.corebot.store.partition
 import com.gatehill.corebot.test.KMySQLContainer
 import org.amshove.kluent.`should be null`
 import org.amshove.kluent.`should equal`
+import org.amshove.kluent.`should not be null`
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.given
 import org.jetbrains.spek.api.dsl.it
@@ -32,12 +33,21 @@ object MysqlDataStoreImplSpec : Spek({
             val partition = store.partition<String, Example>("test")
 
             it("saves objects") {
-                partition["key"] = Example("value")
+                partition["key"] = Example("bar")
             }
 
             it("loads saved objects") {
-                val value = partition["key"]
-                value `should equal` value
+                val savedValue = partition["key"]
+                savedValue.`should not be null`()
+                savedValue!!.foo `should equal` "bar"
+            }
+
+            it("updates existing objects") {
+                partition["key"] = Example("baz")
+
+                val updatedValue = partition["key"]
+                updatedValue.`should not be null`()
+                updatedValue!!.foo `should equal` "baz"
             }
 
             it("removes saved objects") {
@@ -52,4 +62,4 @@ object MysqlDataStoreImplSpec : Spek({
     }
 })
 
-data class Example(val name: String)
+data class Example(val foo: String)